📑 Move template yml types from jtex to myst-templates
