+++ packages/jtex/package.json
@@ -61,6 +61,7 @@
     "js-yaml": "^4.1.0",
     "myst-cli-utils": "^0.0.4",
     "myst-frontmatter": "^0.0.1",
+    "myst-templates": "^0.0.1",
     "node-fetch": "^2.6.7",
     "nunjucks": "^3.2.3",
     "pretty-hrtime": "^1.0.3",
