+++ packages/jtex/src/types.ts
@@ -1,5 +1,5 @@
 import type { ISession as BaseISession } from 'myst-cli-utils';
-import type { Author, Licenses, PageFrontmatter } from 'myst-frontmatter';
+import type { Author, PageFrontmatter } from 'myst-frontmatter';
 import { PAGE_FRONTMATTER_KEYS } from 'myst-frontmatter';
 
 export interface ISession extends BaseISession {
@@ -10,104 +10,6 @@ export type TemplateImports = {
   imports: string[];
   commands: Record<string, string>;
 };
-
-export type TemplatePartDefinition = {
-  id: string;
-  title?: string;
-  description?: string;
-  required?: boolean;
-  plain?: boolean;
-  max_chars?: number;
-  max_words?: number;
-  condition?: {
-    id: string;
-    value?: any;
-  };
-};
-
-export enum TemplateOptionTypes {
-  boolean = 'boolean',
-  string = 'string',
-  choice = 'choice',
-}
-
-export type TemplateDocDefinition = {
-  id: string;
-  title?: string;
-  description?: string;
-  required?: boolean;
-  condition?: {
-    id: string;
-    value?: any;
-  };
-};
-
-export type TemplateOptionDefinition = TemplateDocDefinition & {
-  type: TemplateOptionTypes;
-  default?: any;
-  choices?: string[];
-  max_chars?: number;
-};
-
-export type TemplateStyles = {
-  citation?: 'numerical-only';
-  bibliography?: 'natbib' | 'biblatex';
-};
-
-type TemplateYmlListPartial = {
-  title?: string;
-  description?: string;
-  version?: string;
-  authors?: Author[];
-  license?: Licenses;
-  tags?: string[];
-};
-
-type TemplateYmlPartial = {
-  jtex: 'v1';
-  github?: string;
-  build?: { engine?: string };
-  style?: TemplateStyles;
-  parts?: TemplatePartDefinition[];
-  doc?: TemplateDocDefinition[];
-  options?: TemplateOptionDefinition[];
-  packages?: string[];
-  files?: string[];
-};
-
-type TemplateYmlIdLinks = {
-  id: string;
-  links: {
-    self: string;
-    download: string;
-    thumbnail: string;
-    source?: string;
-  };
-};
-
-/**
- * Type template.yml files are directly validated against
- */
-export type TemplateYml = TemplateYmlPartial &
-  TemplateYmlListPartial & {
-    source?: string;
-    thumbnail?: string;
-  };
-
-/**
- * Type for /template/tex API list response
- */
-export type TemplateYmlListResponse = {
-  items: (TemplateYmlListPartial &
-    TemplateYmlIdLinks & {
-      kind: string;
-    })[];
-};
-
-/**
- * Type for /template/tex/org/name API response
- */
-export type TemplateYmlResponse = TemplateYmlPartial & TemplateYmlListPartial & TemplateYmlIdLinks;
 
 export type ValueAndIndex = {
   value: any;
