+++ packages/myst-templates/src/index.ts
@@ -1,22 +1,104 @@
-import type { Author, License } from 'myst-frontmatter';
+import type { Author, Licenses } from 'myst-frontmatter';
 
 export enum TemplateKind {
   tex = 'tex',
+  docx = 'docx',
 }
 
-export type TemplateDTO = {
+export type TemplatePartDefinition = {
   id: string;
-  kind: TemplateKind;
-  title: string;
-  description: string;
-  authors: Author[];
-  license?: License;
-  tags: string[];
-  version: string;
-  links: {
-    self: string;
-    source: string;
-    thumbnail: string;
-    download: string;
+  title?: string;
+  description?: string;
+  required?: boolean;
+  plain?: boolean;
+  max_chars?: number;
+  max_words?: number;
+  condition?: {
+    id: string;
+    value?: any;
   };
 };
+
+export enum TemplateOptionTypes {
+  boolean = 'boolean',
+  string = 'string',
+  choice = 'choice',
+}
+
+export type TemplateDocDefinition = {
+  id: string;
+  title?: string;
+  description?: string;
+  required?: boolean;
+  condition?: {
+    id: string;
+    value?: any;
+  };
+};
+
+export type TemplateOptionDefinition = TemplateDocDefinition & {
+  type: TemplateOptionTypes;
+  default?: any;
+  choices?: string[];
+  max_chars?: number;
+};
+
+export type TemplateStyles = {
+  citation?: 'numerical-only';
+  bibliography?: 'natbib' | 'biblatex';
+};
+
+type TemplateYmlListPartial = {
+  title?: string;
+  description?: string;
+  version?: string;
+  authors?: Author[];
+  license?: Licenses;
+  tags?: string[];
+};
+
+type TemplateYmlPartial = {
+  jtex: 'v1';
+  github?: string;
+  build?: { engine?: string };
+  style?: TemplateStyles;
+  parts?: TemplatePartDefinition[];
+  doc?: TemplateDocDefinition[];
+  options?: TemplateOptionDefinition[];
+  packages?: string[];
+  files?: string[];
+};
+
+type TemplateYmlIdLinks = {
+  id: string;
+  links: {
+    self: string;
+    download: string;
+    thumbnail: string;
+    source?: string;
+  };
+};
+
+/**
+ * Type template.yml files are directly validated against
+ */
+export type TemplateYml = TemplateYmlPartial &
+  TemplateYmlListPartial & {
+    source?: string;
+    thumbnail?: string;
+  };
+
+/**
+ * Type for /template/tex API list response
+ */
+export type TemplateYmlListResponse = {
+  items: (TemplateYmlListPartial &
+    TemplateYmlIdLinks & {
+      kind: TemplateKind;
+    })[];
+};
+
+/**
+ * Type for /template/tex/org/name API response
+ */
+export type TemplateYmlResponse = TemplateYmlPartial & TemplateYmlListPartial & TemplateYmlIdLinks;
